Jupyter Nextcloud App: Working app
This is a first draft of a Nextcloud App that can embed Jupyter
notebooks from a separate stand alone JupyterHub instance.

There is one configuration key called jupyter_url in oc_appconfig. You
can set this with occ jupyter:set-url <URL to your Jupyter instance>.
You can also get the url which is set with jupyter:get-url.

There is also a possibility to set this from the "Additional settings"
menu.